chore(crate): move all tests to their folder
