+++ src/api/search.rs
@@ -110,86 +110,3 @@ impl ModrinthAPI {
         self.search(query, sort, limit, extended_search).await
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::structs::{projects::ProjectType, search::Facet};
-
-    #[tokio::test]
-    async fn search_project() -> Result<()> {
-        let api = ModrinthAPI::default();
-        let response = api
-            .search(
-                "xaeros",
-                &Sort::Downloads,
-                None,
-                Some(ExtendedSearch {
-                    offset: None,
-                    facets: vec![vec![Facet::ProjectType(ProjectType::Mod)]],
-                }),
-            )
-            .await?;
-
-        let response = response.hits.first().unwrap();
-        let title = response.slug.as_ref();
-
-        assert!(title.is_some());
-        assert_eq!(title.ok_or(0), Ok(&String::from("xaeros-minimap")));
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_fetching_project_with_mut() -> Result<()> {
-        let api = ModrinthAPI::default();
-        let response = api
-            .search(
-                "xaeros",
-                &Sort::Downloads,
-                None,
-                Some(ExtendedSearch {
-                    offset: None,
-                    facets: vec![vec![Facet::ProjectType(ProjectType::Mod)]],
-                }),
-            )
-            .await?;
-
-        let response = response.hits.first().unwrap();
-        let title = response.slug.as_ref();
-
-        assert!(title.is_some());
-        assert_eq!(title.ok_or(0), Ok(&String::from("xaeros-minimap")));
-
-        let hit = response.to_owned().fetch_project(&api).await?;
-
-        assert_eq!(hit.slug.as_ref().unwrap(), &String::from("xaeros-minimap"));
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_fetching_project_without_mut() -> Result<()> {
-        let api = ModrinthAPI::default();
-        let response = api
-            .search(
-                "xaeros",
-                &Sort::Downloads,
-                None,
-                Some(ExtendedSearch {
-                    offset: None,
-                    facets: vec![vec![Facet::ProjectType(ProjectType::Mod)]],
-                }),
-            )
-            .await?;
-
-        let response = response.hits.first().unwrap();
-        let title = response.slug.as_ref();
-
-        assert!(title.is_some());
-        assert_eq!(title.ok_or(0), Ok(&String::from("xaeros-minimap")));
-
-        let hit = response.get_full_project(&api).await?;
-
-        assert_eq!(hit.slug, String::from("xaeros-minimap"));
-        Ok(())
-    }
-}
